+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes for Version 0.1.19
 
 - Added a temporary submission disclaimer for publishing data (#17)
+- Data Policy Dialog Buttons are now labelled Accept and Decline (#7)
 
 # Changed for Version 0.1.7